Fine tuned the blocksize for large sizes and small typesizes.
This allows BloscLZ to achieve better compression ratios in some
scenarios (the ultimate reason for this is that hash_log in BloscLZ
cannot be larger than 15).  However, in other scenarios it achieves
lower compression rations (for example, the bench.c benchmark :-/).

Anyway, I think that the new behaviour is less surprise-prone (see
ticket #7).  Closes #7.



git-svn-id: http://blosc.pytables.org/svn/blosc/trunk@295 5b11b6af-1e7b-4319-8082-0aacbd7ffead