fix codec enumeration (in README)
Upon further inspection of the Blosc source code the changes made to the header
README in 8707b4c15b9ee2d553937c17a3c4e78c7fd87738 must be reverted, my bad. As
it turns out, there is now one set of integers used internally by blosc as
arguments to compress and another that is actually used to identify the codec
in the header. I presume that since LZ4 and LZ4HC have the same decompressor,
the same integer (namely ``1``) is used for both. Updated the docs to account
for that.

Also, it turns out that the field that was previously used for the version of
the blosclz is now more general and stores the version of the codec used.
Updated the README to reflect that change.