Implemented extensible fat-binary support.
Implemented extensible fat-binary support which allows the hardware-
accelerated shuffle implementation to be dynamically selected at
run-time based on the CPU's supported features. This avoids repeatedly
checking the CPU's supported features (which is a serializing operation
on x86-compatible hardware); instead, hardware support is detected once
and cached, so the shuffle functions can then dynamically dispatch to
the appropriate implementation from then on.
This new fat-binary support also allows blosc to be compiled on older
hardware with support for newer hardware (e.g., AVX2) as long as the
compiler supports the newer hardware.
This implementation is structured to make it easy to add hardware-
acceleration support for new platforms (e.g., ARM NEON)

Finally, the compiled blosc library now exports the normal shuffle/
unshuffle functions (the ones that dynamically dispatch to the
accelerated implementations) and any accelerated shuffle/unshuffle
functions (e.g., shuffle_sse2(), unshuffle_sse2()). This is to allow
the shuffle/unshuffle implementations to be tested individually, and
also makes it easier to implement benchmarks to compare the various
implementations to each other.