Implement x86-compatible feature-detection support.
Implemented x86-compatible feature-detection support that works on
GCC < 4.8, Clang, and MSVC.
This removes the dependence on the __builtin_cpu_supports intrinsic
which was only available in GCC 4.8 and newer; that intrinsic is also
flawed in that it doesn't check the value of the extended control
register (XCR0) which indicates whether the OS is able to restore the
state of the SSE and AVX registers. When running an older OS on newer
hardware, trying to use registers unsupported by the OS could cause
the process or system to crash.