Implemented "peeled" AVX2 shuffle.
Modified the AVX2 shuffle implementation to use the same approach as the
'peeled' SSE2 implementation. This allows the AVX2 shuffle to be used
even when the number of elements in the buffer is not a multiple of the
vector size (in this case, 32).

Note: this commit also fixes an addressing bug in shuffle2_avx2 which
caused it to store the shuffled data incorrectly for certain input
buffer sizes.