test was susceptible to an off-by-one error
In case we try to compress the empty string, it is fine if the output
buffer is exactly BLOSC_MAX_OVERHEAD bytes long. Hence we modify the
check to make the output buffer one byte less than BLOSC_MAX_OVERHEAD.
It is now somewhat redundent to check for the output buffer being two
bytes less than BLOSC_MAX_OVERHEAD, but we keep it anyway since it
doesn't hurt.