Use vendored `stdint.h` for older Visual Studio
On older versions of Visual Studio, `stdint.h` is not included. In these
cases, we need to use our vendored copy of `stdint.h` as a stand-in. For
newer versions of Visual Studio or other compilers, `stdint.h` is
generally included. So this workaround is unneeded in those cases and we
handle them accordingly as well.