Add blosc_internal prefix to _xgetbv and __cpuid definitions in shuffle.c
This fixes the build for clang with -march=haswell, where including
immintrin.h includes xsaveintrin.h, which defines _xgetbv.  This
change additional renames __cpuid to prevent possible future breakages
and for conformance with the ISO C standard which reserves all global
names beginning with an underscore for use by the compiler.