MAINT: don't pass NULL to fprintf
If `clibcode_to_clibname` fails, it returns NULL. This result is directly fed
into a %s format in `fprintf`. A null pointer argument for %s is undefined
behavior, but glibc will print the string `"(null)"`.

gcc-9 reports a warning because of this:

```
In function ‘fprintf’,
    inlined from ‘blosc_c’ at /home/joe/projects/c/c-blosc/blosc/blosc.c:704:7:
/usr/include/bits/stdio2.h:100:10: warning: ‘%s’ directive argument is null [-Wformat-overflow=]
  100 |   return __fprintf_chk (__stream, __USE_FORTIFY_LEVEL - 1, __fmt,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  101 |    __va_arg_pack ());
      |    ~~~~~~~~~~~~~~~~~
```

This patch explicitly checks for a NULL result from `clibcode_to_clibname` and
sets the string to "(null)". Given that surrounding code path is unlikely, the
extra NULL check should not have a meaningful performance impact.