Always use inline assembly for _xgetbv on GCC/Clang
This reverts the change to blosc/shuffle.c in
0b8608b5d12ca96d4ac4525b36a62062b2286fd4.

The _xgetbv intrinsic defined by GCC and Clang can only be used when
built with support for the xsave target feature.  Therefore,
attempting to use it from blosc_get_cpu_features results in a compiler
error:

error: inlining failed in call to 'always_inline' '_xgetbv': target specific option mismatch

On most platforms, the intrinsic was not being used because
immintrin.h was not getting transitively included in shuffle.c.  However, on
mingw64 it does get transitively included in shuffle.c, which leads to
the error.

In order to be able to use the intrinsic, it would need to be used
from a separate function that is built with xsave support.  That is
more complicated than just using the inline assembly version.